# clientGenerators/go/openapi/model_error.go
/*
 * Dog Management API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi
// Error struct for Error
type Error struct {
	// The HTTP status code generated by the origin server for this occurrence of the problem
	Status float32 `json:"status,omitempty"`
	// A short, human-readable summary of the problem type
	Title string `json:"title,omitempty"`
}