# openAPI/example3/spec.yaml
openapi: 3.0.2
info:
  title: Dog Management API
  version: 1.0.0
paths:
  '/dogs/{dogId}':
    parameters:
      - name: dogId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getDog
      summary: Fetch dog information by the dog's name
      responses:
        '200':
          description: Successfully returns dog information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
        '404':
          description: Dog not found error
          content:
            application/json:
              example:
                status: 404
                title: Not Found
              schema:
                $ref: '#/components/schemas/Error'
  /dogs:
    post:
      operationId: createDog
      summary: Create new dog
      requestBody:
        description: The dog to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dog'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dog'
          description: Successful created new dog
        '400':
          content:
            application/json:
              example:
                status: 400
                title: Bad Request
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
components:
  schemas:
    Error:
      type: object
      properties:
        status:
          type: number
          maximum: 599
          minimum: 100
          description: The HTTP status code generated by the origin server for this occurrence of the problem
        title:
          type: string
          description: 'A short, human-readable summary of the problem type'
    Dog:
      type: object
      properties:
        id:
          type: string
          readOnly: true
          description: The id of the dog.
          example: sir-barks-a-lot
        name:
          type: string
          description: The name of the dog.
          example: Sir Barks-a-Lot
        age:
          type: number
          description: The age of the dog.
          minimum: 6
          example: 7
      required:
        - id
        - name
        - age